refactor logic back to recordButton
