+++ ui/src/components/RecordButton.js
@@ -17,6 +17,10 @@
 import {LitElement, html} from 'lit-element';
 import {Icon} from '@material/mwc-icon';
 import {CollectionStates} from '../utils/CollectionStatesEnum';
+import {dispatchErrorToast} from '../utils/ToastUtils.js';
+import {QualityControl} from '../utils/QualityControl';
+import {AudioRecorder} from '../utils/AudioRecorder';
+import {UtteranceApiService} from '../utils/UtteranceApiService';
 
 import style from '../styles/components/RecordButton.css.js';
 
@@ -40,10 +44,98 @@ export class RecordButton extends LitElement {
     constructor() {
         super();
         this.disabled = false;
+        this.audioRecorder = new AudioRecorder();
+        this.utteranceService = new UtteranceApiService();
+        this.finishedAudio = undefined;
     }
 
-    updated() {
+    updated(changedProperties) {
+        this.handleWaveCanvas();
         this.determineButtonAttributes();
+    }
+
+    /**
+     * Handles logic to start recording audio.
+     */
+    async handleStartRecording() {
+        // attempt to init; check for browser permission
+        try {
+            await this.audioRecorder.initRecorder();
+        } catch (e) {
+            dispatchErrorToast(
+                this,
+                `Microphone access is currently blocked for this site. 
+                    To unblock, please navigate to chrome://settings/content/microphone 
+                    and remove this site from the 'Block' section.`
+            );
+            return;
+        }
+
+        // start recording
+        if (!this.audioRecorder.startRecording()) {
+            dispatchErrorToast(this, 'Failed to start recording.');
+            return;
+        }
+
+        // Set to recording state
+        this.dispatchCollectionState(CollectionStates.RECORDING);
+
+        this.audioStream = this.audioRecorder.stream;
+        this.context = new (window.AudioContext || window.webkitAudioContext)();
+    }
+
+    /**
+     * Handles logic to end recording and progress to before_recording state
+     */
+    async handleEndRecording() {
+        // Capture audio into variable
+        let audio;
+
+        try {
+            audio = await this.audioRecorder.stopRecording();
+        } catch (e) {
+            dispatchErrorToast(
+                this,
+                `Could not record successfully; ${e.name}: ${e.message}`
+            );
+        }
+
+        // Do auto qc checks
+        const qualityCheck = new QualityControl(this.context, audio.blob);
+        const qualityResult = await qualityCheck.isQualitySound();
+        if (!qualityResult.success) {
+            // If qc failed, pivot to QC error collection state
+            this.qcError = 'qc error';
+            this.dispatchCollectionState(CollectionStates.NOT_RECORDING);
+            return;
+        }
+
+        // Checks finished; transition to before_upload state
+        this.dispatchCollectionState(CollectionStates.BEFORE_UPLOAD);
+
+        this.finishedAudio = audio;
+    }
+
+    /**
+     * Handles logic to upload recording and trigger transition to next prompt.
+     */
+    async handleUploadRecording() {
+        // Attempt to upload it
+        if (this.finishedAudio.recordingUrl) {
+            try {
+                const resp = await this.utteranceService.saveAudio(
+                    this.finishedAudio
+                );
+
+                if (!resp) throw new Error();
+            } catch (e) {
+                // If upload failed, pivot to upload error collection state
+                this.dispatchCollectionState(CollectionStates.UPLOAD_ERROR);
+            }
+        }
+
+        // Dispatch transition to next prompt.
+        this.dispatchCollectionState(CollectionStates.TRANSITIONING);
     }
 
     /**
@@ -52,22 +144,22 @@ export class RecordButton extends LitElement {
     async handleButtonClick() {
         // not_recording to recording
         if (this.collectionState === CollectionStates.NOT_RECORDING) {
-            this.dispatchCollectionState(CollectionStates.RECORDING);
+            this.handleStartRecording();
         }
 
         // recording to before_upload
         else if (this.collectionState === CollectionStates.RECORDING) {
-            this.dispatchCollectionState(CollectionStates.BEFORE_UPLOAD);
+            this.handleEndRecording();
         }
 
         // before_upload to transition
         else if (this.collectionState === CollectionStates.BEFORE_UPLOAD) {
-            this.dispatchCollectionState(CollectionStates.TRANSITIONING);
+            this.handleUploadRecording();
         }
     }
 
     dispatchCollectionState(newState) {
-        const event = new CustomEvent('record-button-update', {
+        const event = new CustomEvent('update-collection-state', {
             detail: {
                 state: newState,
             },
@@ -75,6 +167,23 @@ export class RecordButton extends LitElement {
             composed: true,
         });
 
+        this.dispatchEvent(event);
+    }
+
+    /**
+     * Emits an event that causes the application to render a sound
+     * wave that corresponds to the current audio stream.
+     */
+    handleWaveCanvas() {
+        const event = new CustomEvent('update-wave', {
+            detail: {
+                audioStream: this.audioStream,
+                context: this.context,
+            },
+
+            bubbles: true,
+            composed: true,
+        });
         this.dispatchEvent(event);
     }
